add JogoController e metodos
